requirements.txt
torch
torchvision
pydicom
opencv-python
numpy
pandas
scikit-learn
matplotlib
tqdm
pillow
grad-cam
albumentations
Flask
tensorflow
